requirements.txt
celery[redis]
Django
djangorestframework
django-celery-results
pandas
psycopg2-binary